added frontend to news
