+++ my-app/src/app/(pages)/dashboard/components/news/News.css
@@ -1,268 +1,82 @@
-.news-component {
-    padding: 2rem;
-}
-
-.news-header {
+.newsupload-component {
+    width: 100%;
     display: flex;
-    justify-content: space-between;
+    flex-direction: column;
     align-items: center;
-    margin-bottom: 2rem;
+    justify-content: center;
 }
 
-.news-header h1 {
-    font-size: 2rem;
-    color: #333;
-    margin: 0;
-}
-
-.news-filters {
+.newsupload-component-heading {
+    width: 100%;
     display: flex;
-    gap: 1rem;
-}
-
-.search-input {
-    padding: 0.5rem;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 0.875rem;
-    width: 250px;
-}
-
-/* News Grid Styles */
-.news-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-    gap: 2rem;
-}
-
-.no-news {
-    grid-column: 1 / -1;
-    text-align: center;
-    padding: 3rem;
-    background-color: #f5f5f5;
-    border-radius: 8px;
-}
-
-.news-card {
-    background-color: white;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    overflow: hidden;
-    transition: transform 0.2s;
-}
-
-.news-card:hover {
-    transform: translateY(-4px);
-}
-
-.news-image {
-    width: 100%;
-    height: 200px;
-    overflow: hidden;
-}
-
-.news-image img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.news-content {
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: center;
     padding: 1.5rem;
+    border-radius: 0.75rem;
+    border: 2px solid #34d399;
 }
-
-.news-content h3 {
-    margin: 0 0 1rem 0;
-    color: #333;
-    font-size: 1.25rem;
-}
-
-.news-content p {
-    margin: 0 0 1.5rem 0;
-    color: #666;
-    line-height: 1.5;
-    display: -webkit-box;
-    -webkit-line-clamp: 3;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-}
-
-.news-actions {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    border-top: 1px solid #eee;
-    padding-top: 1rem;
-}
-
-.action-buttons {
-    display: flex;
-    gap: 0.5rem;
-}
-
-.read-more-btn {
-    text-decoration: none;
-    color: #2196F3;
-    font-weight: 500;
-    font-size: 0.875rem;
-}
-
-.read-more-btn:hover {
-    text-decoration: underline;
-}
-
-/* Upload Form Styles */
-.upload-form-container {
-    background-color: white;
-    border-radius: 8px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    padding: 2rem;
-    max-width: 600px;
-    margin: 0 auto;
-}
-
-.upload-form {
-    display: flex;
-    flex-direction: column;
-    gap: 1.5rem;
-}
-
-.form-group {
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-}
-
-.form-group label {
-    font-size: 0.875rem;
-    font-weight: 500;
-    color: #333;
-}
-
-.form-group input,
-.form-group textarea {
-    padding: 0.75rem;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-    font-size: 1rem;
-    font-family: inherit;
-}
-
-.form-group textarea {
-    resize: vertical;
-    min-height: 100px;
-}
-
-.form-group input[type="file"] {
-    padding: 0.5rem 0;
-    border: none;
-}
-
-.image-preview {
-    aspect-ratio: 16/9;
-    border-radius: 4px;
-    overflow: hidden;
-    margin-top: -0.5rem;
-}
-
-.image-preview img {
+.newsupload-component-main {
     width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.form-actions {
+    padding: 2rem;
+    margin-top: 1rem;
     display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.newsupload-component-main form{
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    /* background-color: red; */
+    padding: 1rem;
     gap: 1rem;
-    margin-top: 1rem;
+}
+.newsform-group {
+    width: 60%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    /* border: 1px solid #d1d5db; */
+    /* border-radius: 0.5rem; */
+    /* background-color: red; */
+
+}
+.newsform-group label {
+    width: 50%;
+    font-weight: 600;
+    color: #374151;
 }
 
-.submit-btn, .cancel-btn {
-    padding: 0.75rem 1.5rem;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
+.newsform-group input {
+    width: 50%;
+    padding: 0.75rem 1rem;
+    border-radius: 0.5rem;
+    border: 1px solid #d1d5db;
+    background-color: white;
     font-size: 1rem;
-    transition: background-color 0.2s;
 }
 
-.submit-btn {
-    background-color: #4CAF50;
-    color: white;
-    flex: 2;
+.newsform-group input:focus {
+    outline: none;
+    border-color: #34d399;
+    box-shadow: 0 0 0 3px rgba(52, 211, 153, 0.2);
 }
 
-.submit-btn:hover {
-    background-color: #45a049;
+.newsform-group input:hover {
+    border-color: #9ca3af;
+}
+.newsform-group-button button {
+    width: 100%;
+    padding: 0.75rem 1rem;
+    border-radius: 0.5rem;
+    border: 1px solid #d1d5db;
+    background-color: #34d399;
 }
 
-.cancel-btn {
-    background-color: #f5f5f5;
-    color: #333;
-    flex: 1;
-}
 
-.cancel-btn:hover {
-    background-color: #e0e0e0;
-}
 
-/* Common Button Styles */
-.edit-btn, .delete-btn {
-    padding: 0.25rem 0.5rem;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-    font-size: 0.75rem;
-    transition: background-color 0.2s;
-}
 
-.edit-btn {
-    background-color: #2196F3;
-    color: white;
-}
 
-.edit-btn:hover {
-    background-color: #1976D2;
-}
-
-.delete-btn {
-    background-color: #f44336;
-    color: white;
-}
-
-.delete-btn:hover {
-    background-color: #d32f2f;
-}
-
-.message {
-    margin-top: 1rem;
-    padding: 10px;
-    border-radius: 4px;
-    text-align: center;
-}
-
-.message.success {
-    background-color: #d4edda;
-    color: #155724;
-    border: 1px solid #c3e6cb;
-}
-
-.message.error {
-    background-color: #f8d7da;
-    color: #721c24;
-    border: 1px solid #f5c6cb;
-}
-
-.delete-button {
-    background-color: #ff4444;
-    color: white;
-    border: none;
-    padding: 5px 10px;
-    border-radius: 4px;
-    cursor: pointer;
-    margin-top: 10px;
-}
-
-.delete-button:hover {
-    background-color: #cc0000;
-}
