+++ my-app/src/app/(pages)/dashboard/components/news/NewsUpload.js
@@ -1,7 +1,12 @@
-import { useState } from 'react';
+import { useState , useEffect} from 'react';
 import './News.css';
 
 const NewsUpload = () => {
+    const [news, setNews] = useState([]);
+    const [searchTerm, setSearchTerm] = useState('');
+    useEffect(() => {
+        fetchNews();
+    }, []);
     const [formData, setFormData] = useState({
         title: '',
         description: '',
@@ -39,10 +44,41 @@ const NewsUpload = () => {
         setFormData({ ...formData, [e.target.name]: e.target.value });
     };
 
+    
+    const fetchNews = async () => {
+        try {
+            const response = await fetch('/api/dashboard/news');
+            if (response.ok) {
+                const data = await response.json();
+                setNews(data);
+            }
+        } catch (error) {
+            console.error('Error fetching news:', error);
+        }
+    };
+
+    const handleDelete = async (articleId) => {
+        try {
+            const response = await fetch(`/api/dashboard/news/${articleId}`, {
+                method: 'DELETE',
+            });
+            
+            if (response.ok) {
+                setNews(news.filter(article => article.id !== articleId));
+            }
+        } catch (error) {
+            console.error('Error deleting news:', error);
+        }
+    };
+
+    const filteredNews = news.filter(item =>
+        item.title.toLowerCase().includes(searchTerm.toLowerCase()) ||
+        item.description.toLowerCase().includes(searchTerm.toLowerCase())
+    );
     return (
         <div className="newsupload-component">
             <div className="newsupload-component-heading">
-                <h2>Instructions to update club details</h2>
+                <h2>Instructions to Upload News Articles</h2>
                 <p>Resize Image: Use the provided Canva link to resize the image to the optimal dimensions for website display.
                 <a href="https://www.canva.com/design/DAGVOb7x6hg/H4_YD4-t9s5ZT-iXhWKKjg/view?utm_content=DAGVOb7x6hg&utm_campaign=designshare&utm_medium=link&utm_source=publishsharelink&mode=preview"> Canva Link</a></p>
                 <p>After resizing the image, download it and upload it to a storage service like <a href="https://www.imghippo.com">ImgHippo</a> Please refer to the <a href="https://firebasestorage.googleapis.com/v0/b/sacwebsite-8d0b5.appspot.com/o/Video_Tutorial.mp4?alt=media&token=a9487ecb-40aa-423a-bf20-26150128b7f5">Video Tutorial</a></p>
@@ -89,7 +125,9 @@ const NewsUpload = () => {
                             required
                         />
                     </div>
-                    <button type="submit">Upload News</button>
+                    <div className="newsform-group-button">
+                        <button type="submit">Upload News</button>
+                    </div>
                 </form>
                 {message && <p className="message">{message}</p>}
             </div>
